Find index of minimum value in a subarray
Finish writing the function indexOfMinimum, which takes an array and a
number startIndex, and returns the index of the smallest value that
occurs with index startIndex or greater. If this smallest value occurs
more than once in this range, then return the index of the leftmost
occurrence within this range. Once implemented, uncomment the
Program.assertEqual() at the bottom to verify that the test assertion
passes.

Test it out!
Hint (What's this?)

Add two more checks using Program.assertEqual() to make sure your
function consistently finds the correct answer.