selection sort
Selection sort loops over positions in the array. For each position, it
finds the index of the minimum value in the subarray starting at that
position. Then it swaps the values at the position and at the minimum
index. Write selection sort, making use of the swap and indexOfMinimum
functions.

Once implemented, uncomment the Program.assertEqual() at the bottom to
verify that the test assertion passes.

Test it out!
Hint (What's this?)

Add another check using Program.assertEqual() to make sure your sorting
algorithm consistently returns the correct order. Try it on arrays with
negative numbers or 0 to make sure it still works in those cases.