Insert value before an index
Fix the insert function to correctly insert the given value into the
array.

The insert function takes three parameters as inputs: array, rightIndex,
and value.

Before the insert function is called:
* the elements from array[0] to array[rightIndex] are sorted in
ascending order.

After calling the insert function:
* value and the elements that were previously in array[0] to
array[rightIndex], should be sorted in ascending order and stored in the
elements from array[0] to array[rightIndex+1].

In order to do this, the insert function will need to make room for
value by moving items that are greater than value to the right. It
should start at rightIndex, and stop when it finds an item that is less
than or equal to value, or when it reaches the beginning of the array.
Once the function has made room for value, it can write value to the
array.

Although there are many ways to write this function, you should write it
in a way that is consistent with the hint code.