Implement insertion sort
Insertion sort loops over items in the array, inserting each new item
into the subarray before the new item.

Write insertion sort, making use of the insert function that was written
in the previous challenge.

Once implemented, uncomment the Program.assertEqual() at the bottom to
verify that the test assertion passes.

Make up another array and try sorting it. Maybe one with negative
numbers or 0 in it?