Iterative factorial
Finish the provided factorial function, so that it returns the value n!.
Your code should use a for loop to compute the product 1 * 2 * 3 * ... *
n. If you write the code carefully, you won't need a special case for
when n equals 0.

Once implemented, uncomment the Program.assertEqual() statements at the
bottom to verify that the test assertions pass.