Recursive case
Write the recursive case:
Call factorial but now with a parameter of n - 1, multiply the result by
n, and return the result of the multiplication. You can do this all in
one line of code.

You should assume that n is postive.

Make sure to uncomment the Program.assertEqual() for factorial(5) at the
bottom to verify that it's returning the expected value.