Challenge: is a string a palindrome?
Base case #1

In this challenge, you'll make it so the palindrome() function returns
true if the provided string is a palindrome, and false otherwise.

Start by implementing the first base case:
if the length of the string is 0 or 1, isPalindrome() should return
true.

Once implemented, uncomment the first Program.assertEqual() for isP
alindrome("a") at the bottom to verify that the test assertion passes.

Base case #1

In this challenge, you'll make it so the palindrome() function returns
true if the provided string is a palindrome, and false otherwise.

Start by implementing the first base case:
if the length of the string is 0 or 1, isPalindrome() should return
true.

Once implemented, uncomment the first Program.assertEqual() for
isPalindrome("a") at the bottom to verify that the test assertion
passes.

Recursive case

Finally, write the recursive case.

Use the provided function middleCharacters to remove the first and last
characters from the string.

Check that the console output from the checkPalindrome function is
correct.

Once implemented, uncomment the Program.assertEqual() for
isPalindrome("rotor") at the bottom to verify that the test assertion
passes.