Challenge: Recursive powers
Base case

Write a recursive function power(x, n) that returns the value of x^n.
(assume that n is an integer)

Start by writing the base case.

Once implemented, uncomment the relevant displayPower() to see how the
result is computed, and uncomment the relevant Program.assertEqual() to
make sure the test passes.

Recursive case: n is odd

There are three recursive cases. In this step, write the recursive case
for which n is odd.

You should use the provided function isOdd() to check if n is odd.

Once implemented, uncomment the relevant displayPower() to see how the
result is computed, and uncomment the relevant Program.assertEqual() to
make sure the test passes.

Recursive case: n is even

In this step, write the recursive case for which n is even.

You should use the provided function isEven() to check if n is even.

Once implemented, uncomment the relevant displayPower() to see how the
result is computed, and uncomment the relevant Program.assertEqual() to
make sure the test passes.

Recursive case: n is negative

There are three recursive cases. In this step, write the recursive case
for which n is negative.

Compute x raised to -n recursively, and return the reciprocal of that
number.

Once implemented, uncomment the relevant displayPower() to see how the
result is computed, and uncomment the relevant Program.assertEqual() to
make sure the test passes.