Challenge: Solve Hanoi recursively
Solve the base case

In this challenge, you will solve the towers of Hanoi problem for five
disks, by writing a recursive function solveHanoi that will solve Hanoi
for any positive number of disks.

A call to solveHanoi(numDisks,fromPeg,toPeg) should move numDisks disks
from the peg fromPeg to the peg toPeg.

Start by implementing the base case of zero disks.

Recursively move disks to the spare peg

Make a recursive function call to move the disks sitting on top of the
bottom disk on the fromPeg to the spare peg, i.e. move (numDisks - 1)
disks to the spare peg.

You can find the spare peg by using the getSparePeg function.

A call to hanoi.getSparePeg(peg1,peg2) returns the remaining peg that
isn't peg1 or peg2.

Move one disk to the goal peg.

You have exposed the bottom disk. Move it to the target peg.

This function call is not recursive; don't call solveHanoi for this
step.

A call to hanoi.moveDisk(peg1,peg2) moves the top disk on peg1 to peg2.

Recursively move disks from the spare peg to the target peg.

Make a recursive function call to move the disks from the spare peg back
on top of the bottom disk, i.e. move (numDisks - 1) disks from the spare
peg to the target peg.

Test it out.

Uncomment the call to solveHanoi and the Program.assertEqual statement
below your solveHanoi function, and press the Restart button to test out
your solveHanoi function.

Every time you call moveDisk, the move will be recorded in the console
window. If your function attempts to make an invalid move, the moveDisk
function will print an error message in the console and block further
calls to moveDisk. To call moveDisk after you have made an invalid move,
you will need to press the 'Restart' button.

A call to hanoi.isSolved(toPeg) returns true if all disks are on toPeg
and no invalid moves have been used.