Challenge: Implement merge sort
Implement merge sort

The mergeSort function should recursively sort the subarray array[p..r].
To remind you of the algorithm: if the subarray has size 0 or 1, then
it's already sorted, and so nothing needs to be done. Otherwise, merge
sort uses divide-and-conquer to sort the subarray. Assume that the call
merge(array, p, q, r) performs the merging. Once implemented, uncomment
the Program.assertEqual() at the bottom to verify that the test
assertion passes.