Implement quicksort
The quicksort function should recursively sort the subarray array[p..r].
If the subarray has size 0 or 1, then it's already sorted, and so
nothing needs to be done. Otherwise, quicksort uses divide-and-conquer
to sort the subarray. The divide step should partition the array, the
conquer step should recursively quicksort the partitioned subarrays, and
the combine step should do nothing. Once implemented, uncomment the
Program.assertEqual() at the bottom to verify that the test assertion
passes.